Move name and cookie to `protocol` package.
