+++ agent.go
@@ -2,6 +2,7 @@ package elevengo
 
 import (
 	"github.com/deadblue/elevengo/internal/impl"
+	"github.com/deadblue/elevengo/internal/protocol"
 	"github.com/deadblue/elevengo/lowlevel/api"
 	"github.com/deadblue/elevengo/lowlevel/client"
 	"github.com/deadblue/elevengo/option"
@@ -42,7 +43,7 @@ func New(options ...option.AgentOption) *Agent {
 		llc: impl.NewClient(hc, cdMin, cdMax),
 	}
 	agent.uh.AppVer, _ = agent.getAppVersion()
-	agent.llc.SetUserAgent(api.MakeUserAgent(name, agent.uh.AppVer))
+	agent.llc.SetUserAgent(protocol.MakeUserAgent(name, agent.uh.AppVer))
 
 	return agent
 }
