+++ internal/protocol/name.go
@@ -1,4 +1,4 @@
-package api
+package protocol
 
 import "fmt"
